requirements.txt
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aistudio-sdk==0.3.8
albucore==0.0.36
albumentations==1.4.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrs==25.4.0
bce-python-sdk==0.9.59
beautifulsoup4==4.14.3
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
comiq==0.1.3
contourpy==1.3.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
cython==3.2.4
datasets==4.4.2
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
easyocr==1.7.2
einops==0.8.1
et-xmlfile==2.0.0
evaluate==0.4.6
fastapi==0.127.1
ffmpy==1.0.0
filelock==3.20.3
fire==0.7.1
flash-attn==2.8.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
fugashi==1.5.2
future==1.0.0
googletrans==4.0.0rc1
gradio==6.2.0
gradio-client==2.0.2
groovy==0.1.2
h11==0.16.0
h2==3.2.0
hatchling==1.28.0
hf-xet==1.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.3.2
hyperframe==5.2.0
idna==3.11
imageio==2.37.2
imagesize==1.4.1
jaconv==0.4.1
jinja2==3.1.6
jiter==0.12.0
jiwer==4.0.0
joblib==1.5.3
kintera==1.2.7
kiwisolver==1.4.9
lazy-loader==0.4
lightning-utilities==0.15.2
llama-cpp-python==0.3.16
lmdb==1.7.5
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.33.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
numpy==2.4.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cccl-cu12==12.6.77
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.9.6.50
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.25.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
open-clip-torch==3.2.0
openai==2.14.0
opencc-python-reimplemented==0.1.7
opencv-contrib-python==4.10.0.84
opencv-python==4.8.0.74
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.5
packaging==25.0
paddle==1.2.0
paddleocr==3.3.2
paddlepaddle==2.6.2
paddlepaddle-gpu==3.2.0
paddlex==3.3.13
pandas==2.3.3
pathspec==0.12.1
peft==0.17.1
pillow==12.1.0
pip==24.0
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
premailer==3.10.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydisort==1.4.3
pydub==0.25.1
pygments==2.19.2
pyharp==1.8.5
pyparsing==3.3.1
pypdfium2==5.3.0
pyperclip==1.11.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.2
rapidfuzz==3.14.3
regex==2025.11.3
requests==2.32.5
rfc3986==1.5.0
rich==14.2.0
ruamel-yaml==0.19.1
ruamel-yaml-clib==0.2.15
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
snapy==1.0.3
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.50.0
stringzilla==4.6.0
sympy==1.14.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.7.1
torchmetrics==1.3.1
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.57.1
triton==3.3.1
trove-classifiers==2025.12.1.14
typer==0.21.0
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
ultralytics==8.3.246
ultralytics-thop==2.0.18
unidic-lite==1.0.8
urllib3==2.6.3
uvicorn==0.40.0
wand==0.6.13
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
